HAP-1125 - Add pipeline integration tests
