+++ src/test/java/org/jfrog/hudson/pipeline/integration/PipelineType.java
@@ -1,4 +1,4 @@
-package org.jfrog.hudson.pipeline.integrationtests;
+package org.jfrog.hudson.pipeline.integration;
 
 enum PipelineType {
     SCRIPTED,
